% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_child_vaccine_coverage}
\alias{hl_child_vaccine_coverage}
\title{Percentage of Childhood Vaccine Coverage (2023)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{child_vaccine_coverage_percentage}{Percentage of vaccine coverage. Includes
childhood immunisation statistics (6in1, PCV, rotavirus, MenB, (12 months); 6in1, MMR1, Hib/MenC, PCVb,
MenB Booster (24 months); 6in1, MMR1, Hib/MenC, 4in1, MMR2 (3-5 years); MMR1,
4in1, MMR2 (4-6 years))}
\item{year}{School Year}

...
}
}
\source{
\url{https://publichealthscotland.scot/publications/childhood-immunisation-statistics-scotland/childhood-immunisation-statistics-scotland-quarter-ending-30-june-2024/}
}
\usage{
hl_child_vaccine_coverage
}
\description{
A dataset containing statistics on childhood vaccine coverage in each Council, 2022/23.
}
\keyword{datasets}
