% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_drug_misuse}
\alias{hl_drug_misuse}
\title{Rate of Drug Related Hospital Stays (2021/22)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{drug_related_stays_per_100k}{Rate of hospital stays per 100k population,
using European Age-sex Standardised Rates.A hospital stay, also described as
a continuous inpatient stay or CIS, is defined as an unbroken period of time
that a patient spends as an inpatient or day-case. During a stay a patient
may have numerous episodes as they change consultant, significant facility,
speciality and/or hospital. Stays are counted at the point of discharge, when
all diagnostic information regarding the full stay is available. Mid 2022 population
estimates are used.}
\item{year}{Financial year}

...
}
}
\source{
\url{https://www.opendata.nhs.scot/dataset/drug-related-hospital-statistics-scotland/resource/46f9d70b-8517-4af3-b65e-dbcd13dfa388}

\url{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/population/population-estimates/mid-year-population-estimates/mid-2022#:~:text=Of\%20the\%2032\%20council\%20areas,and\%20Orkney\%20Islands\%20with\%2022\%2C020}
}
\usage{
hl_drug_misuse
}
\description{
A dataset containing statistics on the rate of drug related hospital stays in
each Council, 2021/22.
}
\keyword{datasets}
