% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_low_birth_weight}
\alias{hl_low_birth_weight}
\title{Percentage of Live Births at Term with Low Birth Weight (2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{low_birth_weight_percentage}{Percentage of babies born to term
(>= 37 gestation weeks) at low birth weight. Based on aggregate data from
2020/21 to 2022/23 financial years}
\item{year}{Time period}

...
}
}
\source{
\url{ https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
hl_low_birth_weight
}
\description{
A dataset containing statistics on low birth weights in each Council, 2021.
}
\keyword{datasets}
