% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_overweight_obesity_children}
\alias{hl_overweight_obesity_children}
\title{Percentage of Children Overweight/Obese (2022/23)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{overweight_obese_percentage}{Percentage of children clinically classed
as overweight, obese, and severely obese. BMI is used as the unit of measurement.
Based on children in Primary 1 (aged 4-5).}
\item{year}{School Year period}

...
}
}
\source{
\url{ https://www.opendata.nhs.scot/dataset/primary-1-body-mass-index-bmi-statistics/resource/4a3daa0f-1580-4a59-ac9e-64d9a31a4429}
}
\usage{
hl_overweight_obesity_children
}
\description{
A dataset containing statistics on percentage of children clinically classed
as overweight, obese, and severely obese in each Council, 2022/23.
}
\keyword{datasets}
