% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_physical_activity}
\alias{hl_physical_activity}
\title{Percentage of People Meeting Recommended Activity Levels (2018-22)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{activity_levels_met_percent}{Percentage of people that meet the recommended
activity levels}
\item{year}{Time period}

...
}
}
\source{
\url{https://statistics.gov.scot/resource?uri=http\%3A\%2F\%2Fstatistics.gov.scot\%2Fdata\%2Fscottish-health-survey-local-area-level-data}
}
\usage{
hl_physical_activity
}
\description{
A dataset containing statistics on activity levels in
each Council, 2018-22.
}
\keyword{datasets}
