% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_smoking}
\alias{hl_smoking}
\title{Percentage of Current Smokers (2018-22)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{smoking_percent}{Percentage of people that are current smokers}
\item{year}{Time period}

...
}
}
\source{
\url{ https://statistics.gov.scot/slice?dataset=http\%3A\%2F\%2Fstatistics.gov.scot\%2Fdata\%2Fscottish-health-survey-local-area-level-data&http\%3A\%2F\%2Fpurl.org\%2Flinked-data\%2Fcube\%23measureType=http\%3A\%2F\%2Fstatistics.gov.scot\%2Fdef\%2Fmeasure-properties\%2Fpercent&http\%3A\%2F\%2Fpurl.org\%2Flinked-data\%2Fsdmx\%2F2009\%2Fdimension\%23refPeriod=http\%3A\%2F\%2Freference.data.gov.uk\%2Fid\%2Fgregorian-interval\%2F2016-01-01T00\%3A00\%3A00\%2FP4Y}
}
\usage{
hl_smoking
}
\description{
A dataset containing statistics on smoking status in
each Council, 2018-22.
}
\keyword{datasets}
