% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_vaccine_coverage}
\alias{hl_vaccine_coverage}
\title{Percentage of Vaccine Coverage (2023)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{vaccine_coverage_percentage}{Percentage of vaccine coverage. Includes
HPV immunisation statistics (S1-S4, typically aged 11/12-15/16); teenage
booster immunisation statistics (Td/IPV and MenACWY vaccines for S3-S4,
typically aged 13/14-15/16); and childhood immunisation statistics (includes
6in1, PCV, rotavirus, MenB, (12 months); 6in1, MMR1, Hib/MenC, PCVb,
MenB Booster (24 months); 6in1, MMR1, Hib/MenC, 4in1, MMR2 (3-5 years); MMR1,
4in1, MMR2 (4-6 years))}
\item{year}{School Year}

...
}
}
\source{
\url{https://publichealthscotland.scot/publications/hpv-immunisation-statistics-scotland/hpv-immunisation-statistics-scotland-school-year-202223/}

\url{https://publichealthscotland.scot/publications/teenage-booster-immunisation-statistics-scotland/teenage-booster-immunisation-statistics-scotland-school-year-202223/#:~:text=Pupils\%20who\%20miss\%20the\%20routine,77.3\%25\%2C\%20MenACWY\%2077.4\%25).}

\url{https://publichealthscotland.scot/publications/childhood-immunisation-statistics-scotland/childhood-immunisation-statistics-scotland-quarter-ending-30-june-2024/}
}
\usage{
hl_vaccine_coverage
}
\description{
A dataset containing statistics on vaccine coverage in each Council, 2022/23.
}
\keyword{datasets}
