% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_cancer_screening}
\alias{lives_cancer_screening}
\title{Percentage of Cancer Screening Uptake (2020-2022)}
\format{
A data frame with 32 row and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{cancer_screening_uptake}{Percentage of cancer screening uptake. Only
bowel cancer screening uptake is included here.}
\item{year}{Time period - three year aggregates}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
lives_cancer_screening
}
\description{
A dataset containing statistics on the percentage of cancer screening uptake
in each Council, 2020-2022/
}
\keyword{datasets}
