% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_cardiovascular_conditions}
\alias{lives_cardiovascular_conditions}
\title{Percentage of Discharges per 1,000 for Cardiovascular Conditions (2022/23)}
\format{
A data frame with 32 variables and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{cardiovascular_discharges_per_1k}{Percentage of total cardiovascular discharges per 1k population.
Cardiovascular conditions includes angina, coronary heart disease, heart
attack, and heart failure. Scottish 2022 mid year population estimates were used}
\item{year}{Financial year}

...
}
}
\source{
\url{https://www.opendata.nhs.scot/dataset/scottish-heart-disease-statistics/resource/5379a655-d677-46cf-814d-bc0574ac59e4}

\url{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/population/population-estimates/mid-year-population-estimates/mid-2022#:~:text=Of\%20the\%2032\%20council\%20areas,and\%20Orkney\%20Islands\%20with\%2022\%2C020}
}
\usage{
lives_cardiovascular_conditions
}
\description{
A dataset containing statistics on the percentage of discharges per 1,000 for
cardiovascular conditions in each Council, 2022/23.
}
\keyword{datasets}
