% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_child_poverty}
\alias{lives_child_poverty}
\title{Percentage of Absolute Child Poverty (2022-2023)}
\format{
A data drame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{child_poverty_percentage}{Percentage of children under 16 living in
absolute low incomes families (where a household's income is below 60\% of
the median income in 2010/11, adjusted for inflation).}
\item{year}{Financial Year}

...
}
}
\source{
\url{ https://www.gov.uk/government/statistics/children-in-low-income-families-local-area-statistics-2014-to-2023}
}
\usage{
lives_child_poverty
}
\description{
A dataset containing statistics on the percentage of children (aged under
16) living in absolute low income families in each Council, 2022-23.
}
\keyword{datasets}
