% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_early_years_development}
\alias{lives_early_years_development}
\title{Percentage of Children with Developmental Concerns (2022-23)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{developmental_concerns_percent}{Percentage of developmental reviews with
a concern in one or more domains, out of total number of recorded reviews while
child is 2 and in the cohort.}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.opendata.nhs.scot/dataset/27-30-month-review-statistics/resource/018ba0e1-6562-43bb-82c5-97b6c6cc22d8}
}
\usage{
lives_early_years_development
}
\description{
A dataset containing statistics on early childhood developmental concerns in
each Council, 2022-23.
}
\keyword{datasets}
