% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_high_blood_pressure}
\alias{lives_high_blood_pressure}
\title{Percentage of Adults Diagnosed with High Blood Pressure (2018-2022)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{ high_blood_pressure_percentage}{Percentage of adults (excluding pregnant
women) diagnosed with high blood pressure}
\item{year}{Time period - four year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/sg-scottish-health-survey/}
}
\usage{
lives_high_blood_pressure
}
\description{
A dataset containing statistics on percentage of adults (excluding pregnant
women) diagnosed with high blood pressure, by Council (2018-2022).
}
\keyword{datasets}
