% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_infant_mortality}
\alias{lives_infant_mortality}
\title{Rate of infant mortality (2017-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authroity Code}
\item{infant_mortality_rate_per_1k}{Rate of infants per 1k that have died
under the age of 1}
\item{year}{Time period - five year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
lives_infant_mortality
}
\description{
A dataset containing statistics on the rate of infant mortality in each
Council, 2017-2021.
}
\keyword{datasets}
