% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_national_five_attainment}
\alias{lives_national_five_attainment}
\title{Percentage Grade A-C in Maths and English, Provisional National 5 Attainments (2024)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{national_five_attainment_percent}{Grade A-C percentage in Maths
and English (average of the two subjects taken)}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.sqa.org.uk/sqa/105123.html}
}
\usage{
lives_national_five_attainment
}
\description{
A dataset containing statistics on Provisional National 5 Attainments in Maths and English,
on results day, in each Council, August 2024
}
\keyword{datasets}
