% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_pupil_absence}
\alias{lives_pupil_absence}
\title{Crude Rate of School Exclusion (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{pupil_absence_per_1k}{Crude exclusion rate of pupil absence per 1,000
students}
\item{year}{School year}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
lives_pupil_absence
}
\description{
A dataset containing crude rate of school exclusion rate 1,000 aged 15-19
in each Council, 2019-2021.
}
\keyword{datasets}
