% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_sexual_health}
\alias{lives_sexual_health}
\title{Sexual health Rates per 100k (2019)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{sexual_health_rate_100k}{Rate of chlamydia per 100,000}
\item{year}{Year}

...
}
}
\source{
\url{https://hpspubsrepo.blob.core.windows.net/hps-website/nss/3073/documents/3_genital-chlamydia-gonorrhoea-scotland-2010-2019-tables.xlsx}
}
\usage{
lives_sexual_health
}
\description{
A dataset containing statistics on rates of chlamydia per 100k in each council,
2019.
To note: only chlamydia has been included here as data is missing for
gonorrhoea. England's Health Index examines total number of people tested
for one or more sexually transmitted infections.
}
\keyword{datasets}
