% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_teenage_pregnancy}
\alias{lives_teenage_pregnancy}
\title{Crude Rate of Teenage Pregnancies (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{teenage_pregnancy_per_1k}{Crude rate of teenage pregnancies per 1,000
females aged 15-19}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
lives_teenage_pregnancy
}
\description{
A dataset containing crude rate of teenage pregnancies per 1,000 aged 15-19
in each Council, 2019-2021.
}
\keyword{datasets}
