% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_unemployment}
\alias{lives_unemployment}
\title{Percentage of People in Unemployment (2021-22)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{unemployment_percentage}{Estimated percentage of people that are unemployed.
Unemployment is defined as someone aged 16+ without a job who were available
to start work in the two weeks following their interview and who had either
looked for work in the four weeks prior to interview or were waiting to start
a job they had already obtained}
\item{year}{Financial Year}

...
}
}
\source{
\url{ https://www.ons.gov.uk/employmentandlabourmarket/peoplenotinwork/unemployment/datasets/modelledunemploymentforlocalandunitaryauthoritiesm01}
}
\usage{
lives_unemployment
}
\description{
A dataset containing statistics on estimated unemployment rates in each Council, 2021-22. Estimates based on a model
which utilises the Annual Population Survey estimates of unemployment along with the Claimant Count averaged over 12 months.
}
\keyword{datasets}
