% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_workplace_safety}
\alias{lives_workplace_safety}
\title{Rate of Non-Fatal Workplace Injuries (2022-23)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{non_fatal_injuries_per_100k_employees}{Rate of non-fatal employee
injuries per 100k population}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.hse.gov.uk/statistics/tables/index.htm#riddor}
}
\usage{
lives_workplace_safety
}
\description{
A dataset containing statistics on the rate of non-fatal workplace injuries
per 100,000 in each Council, 2022-23.
}
\keyword{datasets}
