% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_young_people_training}
\alias{lives_young_people_training}
\title{Percentage participation of Young People in Education, Training or Employment (2023)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{young_people_participation_education_training_employment_percentage}{Percentage
of 16-19 year olds in education, training or employment}
\item{year}{Time period}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
lives_young_people_training
}
\description{
A dataset containing data on participation of young people in education, training
or employment, by Council (2023).
}
\keyword{datasets}
