% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_all_mortality}
\alias{people_all_mortality}
\title{Mortality all causes rate per 100k (2023)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{death_rate_per_100k}{age-sex standardised rate for all causes of
mortality per 100k}
\item{year}{Year}

...
}
}
\source{
\url{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/vital-events/deaths/deaths-time-series-data}
}
\usage{
people_all_mortality
}
\description{
A dataset containing statistics from an age-sex standardised rate for all
causes of mortality per 100k, by Council (2023).
}
\keyword{datasets}
