% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_avoidable_deaths}
\alias{people_avoidable_deaths}
\title{Rate of Avoidable Deaths per 100k (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{avoidable_mortality_rate_per_100k}{Age-standardised avoidable mortality rates per
100k}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/vital-events/deaths/avoidable-mortality}
}
\usage{
people_avoidable_deaths
}
\description{
A dataset containing statistics on age-standardised death rates per 100k, by
Council (2019-2021).
}
\keyword{datasets}
