% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_cancer}
\alias{people_cancer}
\title{Cancer Registrations per 100k (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{cancer_registration_rate_per_100k}{Cancer registration per 100k,
standardised by age-sex}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_cancer
}
\description{
A dataset containing statistics on cancer registration per 100k, standardised
by age-sex (2019-2021).
}
\keyword{datasets}
