% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_child_mental_health}
\alias{people_child_mental_health}
\title{Child Mental Wellbeing Score (2013)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{child_mental_wellbeing_score}{Mean mental wellbeing score for S4 pupils.
The Warwick–Edinburgh Mental Wellbeing Scale (WEMWBS) used
here: scoring is between 14-70 with a higher score indicative of greater
positive mental wellbeing}
\item{year}{Time period}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_child_mental_health
}
\description{
A dataset containing secondary school pupils' mental wellbeing scores (2013).
}
\keyword{datasets}
