% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_dementia}
\alias{people_dementia}
\title{Percentage of Long-Stay Nursing Care Residents with Dementia}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{dementia_percentage}{(Percentage of long-stay nursing care residents
medically diagnosed with dementia, by Council. England's Health Index
dementia indicator collects data on the weighted number of people answering
yes to Alzheimer's disease or other cause of dementia in the national GP
Patient Survey. The only available dementia data at Local Authority level in
Scotland focuses on nursing care residents across all sectors.)}
\item{year}{Time period - census data in the year up to 31/03/2024}

...
}
}
\source{
\url{https://www.opendata.nhs.scot/dataset/care-home-census/resource/9bf418aa-c54d-45d3-8306-023e81f49f60}
}
\usage{
people_dementia
}
\description{
A dataset containing statistics on the percentage of long-stay nursing care
residents medically diagnosed with dementia, by Council (2023-24).
}
\keyword{datasets}
