% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_diabetes}
\alias{people_diabetes}
\title{Crude prevalence of diabetes}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{diabetes_percent}{Crude prevalence of diabetes (all types)}
...
}
}
\source{
\url{https://www.diabetesinscotland.org.uk/wp-content/uploads/2023/10/Scottish-Diabetes-Survey-2022.pdf}
}
\usage{
people_diabetes
}
\description{
Crude prevalence of diabetes
}
\keyword{datasets}
