% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_disability_daily_activities}
\alias{people_disability_daily_activities}
\title{Percentage of People Aged 16-64 with Long-Term Health Problem/Disability
that Limits Daily Activities}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{disability_daily_activities_percentage}{percentage of people aged 16-64
with long-Term health problem/disability that limits daily activities by a
lot or by a little.}
\item{year}{Time period}

...
}
}
\source{
\url{https://statistics.ukdataservice.ac.uk/dataset/scotland-s-census-2022-uv303a-long-term-health-problem-or-disability-by-sex-by-age-20-groups/resource/5ccf8e62-96d1-4b13-8871-39082b0c5f49}
}
\usage{
people_disability_daily_activities
}
\description{
A dataset containing statistics on the percentage of people aged 16-64 with
long-Term health problem/disability that limits daily activities, by Council
(2022).
}
\keyword{datasets}
