% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_healthy_life_expectancy}
\alias{people_healthy_life_expectancy}
\title{Average Healthy Life Expectancy for Men and Women (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{healthy_life_expectancy_male}{Weighted population average healthy life
expectancy for men and women}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_healthy_life_expectancy
}
\description{
A dataset containing statistics on average healthy life expectancy for men
and women, by Council (2019-2021).
}
\keyword{datasets}
