% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_healthy_life_expectancy_men}
\alias{people_healthy_life_expectancy_men}
\title{Average Healthy Life Expectancy for Men (2019-2021)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{healthy_life_expectancy_male}{Average healthy life expectancy for men.
England's Health Index uses a weighted population average. Available Scottish
data only has the average mean healthy life expectancy for men.}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_healthy_life_expectancy_men
}
\description{
A dataset containing statistics on average healthy life expectancy for men,
by Council (2019-2021).
}
\keyword{datasets}
