% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_life_worthwhileness}
\alias{people_life_worthwhileness}
\title{Average Measurement of Life Worthwhileness Out of 10 (2022-23 / Orkney Islands
(2020-21))}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{worthwhile_score_out_of_10}{Average personal ratings on feelings
of life worthwhileness out of 10, 10 - most worthwhile, 1 - least worthwhile}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.ons.gov.uk/datasets/wellbeing-local-authority/editions/time-series/versions/4}
}
\usage{
people_life_worthwhileness
}
\description{
A dataset containing statistics of personal ratings on feelings of life
worthwhileness out of 10, by Council (2022-23) except Orkney Islands (2020-21).
}
\keyword{datasets}
