% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_mental_health_conditions}
\alias{people_mental_health_conditions}
\title{Population Prescribed Drugs for Mental Health Conditions (2021-22)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{mental_health_conditions_percentage}{Percentage of population prescribed
drugs for anxiety, depression or psychosis}
\item{year}{Time period - financial year}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_mental_health_conditions
}
\description{
A dataset containing statistics of population prescribed drugs for mental health
conditions, by Council (2020-21).
}
\keyword{datasets}
