% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_rheumatoid_arthritis}
\alias{people_rheumatoid_arthritis}
\title{Rheumatoid Arthritis prevelance}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{rheumatoid_arthritis_prevelenace}{Prelenave of Rheumatoid Arthritis}
...
}
}
\source{
\url{https://www.versusarthritis.org/policy/resources-for-policy-makers/musculoskeletal-calculator/download-full-msk-calculator-datasets/}
}
\usage{
people_rheumatoid_arthritis
}
\description{
Rheumatoid Arthritis prevelance
}
\keyword{datasets}
