% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_suicides}
\alias{people_suicides}
\title{Deaths from Suicide (2018-2022)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{suicides_per_100k}{Deaths from suicide (16+).
Age standardised rate per 100,000.}
\item{year}{Time period - 5 year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
people_suicides
}
\description{
A dataset containing statistics from deaths by suicide, by Council.
}
\keyword{datasets}
