% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_air_pollution}
\alias{places_air_pollution}
\title{Air pollution}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla21_code}{Local Authority Code}
\item{air_pollution_weighted}{Population-weighted annual mean PM2.5}
...
}
}
\source{
\url{https://uk-air.defra.gov.uk/data/pcm-data}
}
\usage{
places_air_pollution
}
\description{
A dataset containing population-weighted annual mean PM2.5 data for 2023, by
Scottish Council Area. Uses the anthropogenic component.
}
\keyword{datasets}
