% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_gp_appointments}
\alias{places_gp_appointments}
\title{Acceptable GP appointments}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{acceptable_gp_appointments}{The percentage of people who rated their GP
practice as positive for the care provided}
\item{year}{Time period}
...
}
}
\source{
\url{https://www.gov.scot/publications/health-and-care-experience-survey-results-by-geography/}
}
\usage{
places_gp_appointments
}
\description{
A dataset containing the percentage of people who rated their GP practice as
positive for the care provided, by Council (2023/24)
}
\keyword{datasets}
