% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_internet_access}
\alias{places_internet_access}
\title{Percentage of Households with Home Internet Access (2022)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{internet_access_percentage}{percentage of households with home internet
access. Data is missing for Moray and South Ayrshire}
\item{year}{Year}

...
}
}
\source{
\url{https://www.gov.scot/publications/scottish-household-survey-2022-key-findings/documents/}
}
\usage{
places_internet_access
}
\description{
A dataset containing statistics on percentage of households with home internet
access, by Council (2022).
To Note: Data for Moray and South Ayrshire is missing.
}
\keyword{datasets}
