% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_job_training}
\alias{places_job_training}
\title{Percentage of Job Related Training (2023-24)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{job_related_training_perc}{Percentage of all who received job related
training in last 4 wks - aged 16-64}
\item{year}{Time period}

...
}
}
\source{
\url{https://www.nomisweb.co.uk/datasets/apsnew}
}
\usage{
places_job_training
}
\description{
A dataset containing job related training in Scottish councils.
}
\keyword{datasets}
