% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_low_level_crime}
\alias{places_low_level_crime}
\title{Low-level crime (2023-4)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{low_level_crimes_per_10000}{Number of low-level crimes per 10,000 people}
\item{year}{Year}

...
}
}
\source{
\url{https://www.gov.scot/publications/recorded-crime-scotland-2023-24/}
}
\usage{
places_low_level_crime
}
\description{
A dataset containing statistics on 'low-level crime' per 10,000 people in Scottish
Council Areas. England's Health Index defines 'low-level crimes' as bicycle theft
and shoplifting. Of the two, Scotland only publishes shoplifting data. There
are other crimes listed that could potentially count as 'low-level crimes'
but we will avoid making a potentially arbitrary judgement here.
}
\keyword{datasets}
