% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_personal_crime}
\alias{places_personal_crime}
\title{Personal crime (2023-24)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{personal_crimes_per_10000}{Number of personal crimes per 10,000 people}
\item{year}{Year}

...
}
}
\source{
\url{https://www.gov.scot/publications/recorded-crime-scotland-2023-24/}
}
\usage{
places_personal_crime
}
\description{
A dataset containing statistics on 'personal crime' per 10,000 people in Scottish
Council Areas. 'Personal crime' is defined as
\itemize{
\item Non-sexual crimes of violence (includes robbery)
\item Sexual crimes
\item Damage and reckless behaviour
}
}
\keyword{datasets}
