% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_private_outdoor_space}
\alias{places_private_outdoor_space}
\title{Percentage of Addresses with Access to Private Outdoor Space (2020)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{private_outdoor_space_percentage}{Percentage of addresses with access
to private outdoor space}
\item{year}{Year}

...
}
}
\source{
\url{https://www.ons.gov.uk/economy/environmentalaccounts/datasets/accesstogardensandpublicgreenspaceingreatbritain}
}
\usage{
places_private_outdoor_space
}
\description{
A dataset containing statistics on the percentage of addresses with access
to private outdoor space, by Council (2020).
}
\keyword{datasets}
