% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_rough_sleeping}
\alias{places_rough_sleeping}
\title{Rough sleeping (2023-24)}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{rough_sleeping_per_10k}{number of homelessness applications where
at least one member of the household was sleeping rough within the previous
three months, normalised per 100,000 residents.}
\item{year}{Year}

...
}
}
\source{
\url{https://www.gov.scot/publications/homelessness-in-scotland-2023-24/}
}
\usage{
places_rough_sleeping
}
\description{
A dataset containing statistics on number of homelessness applications where
at least one member of the household was sleeping rough within the previous
three months, normalised per 100,000 residents.
}
\keyword{datasets}
