% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_sports_centre_travel_time}
\alias{places_sports_centre_travel_time}
\title{Distance to sports facilities}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla21_code}{Local Authority Code}
\item{median_travel_time}{Median travel time (in minutes) to the nearest
sports centre among all the Intermediate Zones within a Local Authority}
\item{year}{Year the data was last updated}
...
}
}
\source{
\url{https://openstreetmap.org/}
}
\usage{
places_sports_centre_travel_time
}
\description{
A dataset containing the average (median) travel time (in minutes) to the nearest
sports centre in each Council Areas. This is based on the travel time to the
nearest sports centre in each Intermediate Zone within a Council Area. This may
include sports centres located in England.
}
\keyword{datasets}
