% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_traffic_volume}
\alias{places_traffic_volume}
\title{Road traffic volume}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla21_code}{Local Authority Code}
\item{traffic_volume}{}
...
}
}
\source{
\url{https://roadtraffic.dft.gov.uk/}
}
\usage{
places_traffic_volume
}
\description{
A dataset containing volume of road traffic (millions of vehicle miles per
square kilometre) in Scottish Local Authorities.
}
\details{
The raw local authority traffic figures give the total volume (vehicle miles)
of traffic across each local authority for the whole year.
}
\keyword{datasets}
