% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_low_birth_weight}
\alias{hl_low_birth_weight}
\title{Percentage of Babies Born Not at a Healthy Birth Weight (2020-2022)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{ltla19_code}{Local Authority Code}
\item{not_healthy_birth_rate_percentage}{Percentage of babies born not at a healthy birth weight}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://scotland.shinyapps.io/ScotPHO_profiles_tool/}
}
\usage{
hl_low_birth_weight
}
\description{
A dataset containing percentage babies born not at a healthy birth weight.
}
\keyword{datasets}
