% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_sexual_health}
\alias{lives_sexual_health}
\title{Sexual health rates}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{sexual_health_rate_100k}{Rate of chlamydia per 100,000}

...
}
}
\source{
\url{https://hpspubsrepo.blob.core.windows.net/hps-website/nss/3073/documents/3_genital-chlamydia-gonorrhoea-scotland-2010-2019-tables.xlsx}
}
\usage{
lives_sexual_health
}
\description{
Sexual health rates
}
\keyword{datasets}
