% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_road_safety}
\alias{places_road_safety}
\title{Road safety (2022)}
\format{
A data frame with 32 rows and 2 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{road_accident_count_by_area}{Number of people killed or seriously injured, normalised by land area (km^2)}
\item{year}{Year}
...
}
}
\source{
\url{https://statistics.gov.scot/data/road-safety}
}
\usage{
places_road_safety
}
\description{
A dataset containing number of people killed or seriously injured in each
Scottish Council Area in 2022 (latest available data), normalised by the area
in each Council Area (in square km).
}
\keyword{datasets}
